feat: initial working EEPROM implementation
currently mechanisms for synchronization are a bit weird, but work. some
components still need some standardization on value mapping _(e.g.:
should we store the value from 0 to 100 or from 0 to 127? how we hydrate
elements from that state?)_; as well as encoder rehydration are still
not in place (not because they're hard, but because it is late)