feat: initial working EEPROM implementation (#2)
* feat: initial working EEPROM implementation

currently mechanisms for synchronization are a bit weird, but work. some
components still need some standardization on value mapping _(e.g.:
should we store the value from 0 to 100 or from 0 to 127? how we hydrate
elements from that state?)_; as well as encoder rehydration are still
not in place (not because they're hard, but because it is late)

* feat: encoders, updates, feeling good. xymap coming

* feat: add xymap support to eeprom storage

* feat: fairly stable implementation of EEPROM sync

* feat: EEPROM feature complete

* chore: cleanup

* feat: almost stable EEPROM implementation

- had to add a new method to the XSlider component from GUIslice to
  update its view without calling the callback, so that visual syncs
  from the EEPROM state can be done without triggering midi side
  effects. (these will come, just in a different fashion)

- changes the way we do saves to use an interval + look for changes
  between the states when actually writing to EEPROM to avoid wearing
  that memory address. elements dont call __eeprom_save() anymore. that
  is managed by __eeprom_loop() in the loop() function.

  - also changed some signatures to rely only on the
  channel number instead of the reference of the state.

* fix: add missing __eeprom_loop() function

* feat: stable performance

turned off midi thru, which was good for event handling and stopped many
of the weird quirks when syncing from ableton into the device.

fixed the slider implementation to map correctly for all events.

and some refactoring